# internal/httpapi/openapi.yaml
...
openapi: 3.0.3
info:
  title: Sharpline API
  version: "1.0.0"
  summary: Real-time sports odds, play-money wagering, and line history.
  description: |
    Sharpline is a **play-money simulation of a sportsbook**, not a licensed
    sportsbook. No real money moves through this API. There is no deposit, no
    withdrawal, no payment instrument, no identity verification and no
    geolocation gating, and there never will be.

    All balances are play money and are derived by folding a double-entry
    ledger; no balance is stored as a mutable field anywhere in the system.

    Prices originate from an odds provider (or, in development, from a
    synthetic stochastic market maker) and are normalized and priced by the
    ingest and pricer services before they reach this API. Every price object
    reports the provider instant it was observed at so a client can compute
    staleness itself rather than trusting freshness by implication.
  license:
    name: Unlicensed - portfolio project
  contact:
    name: Sharpline
servers:
  - url: /api/v1
    description: |
      The reverse proxy (deploy/proxy/Caddyfile) forwards `/api/*` to this
      service WITHOUT stripping the prefix, so the paths below are absolute
      as written under `/api/v1`. The proxy is the only published port; this
      service is never addressed directly from a browser.
...
security:
  - {}
paths:
...
  /sports:
    get:
      tags: [catalogue]
      operationId: listSports
      summary: Every sport in the catalogue.
      description: |
        Reads Postgres (`sports`). Single page: the row count is the number of
        sports the provider covers, which is single digits.
      responses:
        "200":
          description: The sports, by display name.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SportPage" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }
  /sports/{sportSlug}/leagues:
    get:
      tags: [catalogue]
      operationId: listLeaguesInSport
      summary: The leagues under one sport.
      description: |
        Reads Postgres (`leagues`, index `leagues_sport_idx`). Single page:
        bounded by the catalogue.
      parameters:
        - $ref: "#/components/parameters/SportSlug"
      responses:
        "200":
          description: The leagues, by display name.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LeaguePage" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }
  /books:
    get:
      tags: [catalogue]
      operationId: listBooks
      summary: Every sportsbook whose lines are ingested.
      description: |
        Reads Postgres (`books`). Single page: bounded by the catalogue.

        `kind` and `is_reference` are both returned. A quote from a book of kind
        `synthetic` is a statement about a random number generator and a client
        MUST be able to label it as such; `is_reference` names the sharp book the
        pricer devigs against, which is the baseline every EV number is relative
        to.
      responses:
        "200":
          description: The books, by display name.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BookPage" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }

  # ---------------------------------------------------------------------------
  # Board
  # ---------------------------------------------------------------------------
  /board:
    get:
      tags: [board]
      operationId: getBoard
      summary: The live odds board across every league.
      description: |
        Tradeable events (`scheduled`, `live`, `suspended`) starting before
        `starting_before`, soonest first, with the current best price on each
        event's main markets.

        **Where this reads from.** Events and their market tree come from
        Postgres, which is the source of truth for the catalogue. Prices come
        from the `prices` hypertable through a bounded window
        (`LatestPriceForEachBookOnSelections`), which is a `DISTINCT ON` served
        entirely by `prices_natural_key_idx`. Redis fronts that read as a
        snapshot cache; a miss falls through to Postgres and the answer is
        identical either way, because CLAUDE.md section 3 makes Redis "never
        the source of truth". A suspended market is returned with
        `status: suspended` rather than hidden, because a suspension is usually
        seconds long and dropping the row would make the board flicker.

        **Staleness.** Every price carries `observed_at` (the provider's own
        instant) and `ingested_at` (when this system received it). Their
        difference is the provider-attributable share of the staleness SLO;
        `now() - observed_at` is total staleness. The API does not editorialise
        either number.
      parameters:
        - $ref: "#/components/parameters/StartingBefore"
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/OddsFormatQuery"
        - $ref: "#/components/parameters/BookFilter"
      responses:
        "200":
          description: One page of the board.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BoardPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }
  /leagues/{leagueSlug}/board:
    get:
      tags: [board]
      operationId: getLeagueBoard
      summary: The live odds board for one league.
      description: |
        Identical to `/board` but scoped to a league, and served by a different
        index (`events_league_board_idx`) because a cross-league query cannot use
        an index led by `league_id`.
      parameters:
        - $ref: "#/components/parameters/LeagueSlug"
        - $ref: "#/components/parameters/StartingBefore"
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/OddsFormatQuery"
        - $ref: "#/components/parameters/BookFilter"
      responses:
        "200":
          description: One page of the league's board.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BoardPage" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }
  /events/{eventId}:
    get:
      tags: [catalogue]
      operationId: getEvent
      summary: One event with its full market tree.
      description: |
        The event detail page: the event, its league and sport breadcrumb, every
        market on it, every selection under each market, and the current price
        from every quoting book on every selection.

        Selections are returned in `domain.SelectionRole.DisplayOrder()` --
        home, draw, away, over, under, outright -- which is not the
        lexicographic order of those strings and therefore cannot be produced by
        any SQL `ORDER BY`. The API sorts them so that every client renders the
        same tree.

        Not paginated: an event's market tree is bounded by the event.
      parameters:
        - $ref: "#/components/parameters/EventID"
        - $ref: "#/components/parameters/OddsFormatQuery"
        - $ref: "#/components/parameters/BookFilter"
      responses:
        "200":
          description: The event and its markets.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventDetail" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }
  /markets/{marketId}/prices:
    get:
      tags: [catalogue]
      operationId: compareMarketPrices
      summary: Multi-book comparison for one market.
      description: |
        Every book's current price on every selection of one market, plus the
        best price per selection and the market's overround at each book.

        This is the endpoint behind CLAUDE.md section 6's "multi-book
        comparison". The overround is computed here rather than stored: it is a
        pure function of the quotes returned in the same response
        (`odds.Overround`), so a client can check the arithmetic, and a stored
        copy would be a second truth able to disagree with the prices beside it.

        Not paginated: bounded by (selections on a market) x (books ingested).
      parameters:
        - $ref: "#/components/parameters/MarketID"
        - $ref: "#/components/parameters/OddsFormatQuery"
        - $ref: "#/components/parameters/BookFilter"
      responses:
        "200":
          description: The comparison grid.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MarketComparison" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }

  # ---------------------------------------------------------------------------
  # History
  # ---------------------------------------------------------------------------
  /selections/{selectionId}/history:
    get:
      tags: [history]
      operationId: getSelectionHistory
      summary: Line movement for one selection at one book.
      description: |
        The reason the `prices` hypertable exists. Returns every quote one book
        made on one selection inside a half-open window `[from, to)`, oldest
        first, optionally downsampled.

        **The window is REQUIRED and is not a convenience.** `prices` is
        partitioned on `observed_at` in 12-hour chunks and migration 00004
        installs no retention policy, so the chunk count grows for the life of
        the deployment. A read with no lower bound defeats chunk exclusion and
        makes the planner consult an index on every chunk that has ever existed.
        There is deliberately no unbounded variant to reach for by accident.
        `to` defaults to now; `from` has no default and must be supplied.

        **Downsampling.** `resolution` selects a Timescale `time_bucket` width.
        `raw` returns every stored quote and is what a five-minute window wants;
        anything longer wants a bucket, because a market quoted every two seconds
        for a week is 300k points and no chart can draw them. A bucketed point
        reports `open`, `high`, `low`, `close` and `samples` for the bucket
        rather than a mean: the mean of a line is not a line anyone traded at,
        and open/close is what CLV and steam detection are computed from.

        `max_points` is a guard, not a page size -- a request whose window and
        resolution would exceed it is rejected `422` with the resolution that
        would fit, rather than being silently truncated into a chart that lies
        about where the line ended.
      parameters:
        - $ref: "#/components/parameters/SelectionID"
        - name: book
          in: query
          required: true
          description: |
            The book whose quotes to return, by slug. Required: a chart mixing
            two books' quotes on one series is not a line movement chart, it is
            two of them overlaid, and the caller must say which it wants.
          schema: { $ref: "#/components/schemas/Slug" }
        - name: from
          in: query
          required: true
          description: Inclusive lower bound on `observed_at` (RFC 3339).
          schema: { type: string, format: date-time }
        - name: to
          in: query
          required: false
          description: |
            Exclusive upper bound on `observed_at` (RFC 3339). Defaults to the
            time the request is served. Half-open so adjacent windows tile
            without double-counting the boundary quote.
          schema: { type: string, format: date-time }
        - name: resolution
          in: query
          required: false
          description: |
            Downsampling bucket width. `raw` returns stored quotes unaggregated.
          schema: { $ref: "#/components/schemas/HistoryResolution" }
        - name: max_points
          in: query
          required: false
          description: Reject (422) rather than truncate if the window would exceed this many points.
          schema: { type: integer, format: int32, minimum: 1, maximum: 5000, default: 1000 }
        - $ref: "#/components/parameters/OddsFormatQuery"
      responses:
        "200":
          description: The line movement series.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HistorySeries" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }

  # ---------------------------------------------------------------------------
  # Search
  # ---------------------------------------------------------------------------
...
  /auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: Revoke the presented refresh token's entire family.
      description: |
        Revokes the family with reason `logout`, which ends every session
        descended from that login. Idempotent: logging out an already-revoked or
        unknown family returns `204`, because reporting "that token was already
        dead" is an enumeration oracle and there is nothing a caller can do with
        the distinction.
      security: [{}]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "204":
          description: The family is revoked (or was already).
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
        "500": { $ref: "#/components/responses/InternalError" }

  # ---------------------------------------------------------------------------
  # Account
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A short-lived JWT access token from `/auth/login` or `/auth/refresh`.

        The verifier PINS the signing algorithm and ignores the `alg` header on
        the presented token, so `alg: none` and an algorithm-confusion downgrade
        are both unrepresentable rather than merely rejected. Issuer, audience
        and expiry are all verified. Lifetime is minutes; durable session length
        comes from refresh-token rotation, not from a long-lived access token.
...